adjusted prices for splits. MAE -> MAPE

+++ code/preprocess.py
@@ -1,6 +1,7 @@
-import pandas
+import pandas as pd
 from alpha_vantage.timeseries import TimeSeries
 import time
+import os
 
 api_key = 'A759HSE0DRDPALHC'
 path_to_data = "../data/stocks/"
@@ -23,6 +24,14 @@ def get_data_paths():
         stock_files.append(path)
     return stock_files
     
+def adjust_for_splits(df : pd.DataFrame):
+    """
+    Adjusts stock price based on the stock splits
+    """
+    print(df)
+    print(df['Split Coefficient'].cumprod())
+    
+    return df
 
 def clean_data():
     """
@@ -34,14 +43,19 @@ def clean_data():
                       "2. high": "High",
                       "3. low": "Low",
                       "4. close": "Close",
-                      "5. volume": "Volume"}
+                      "5. adjusted close": "Adjusted Close",
+                      "6. volume": "Volume",
+                      "7. dividend amount": "Dividend Amount",
+                      "8. split coefficient": "Split Coefficient"}
     stocks = get_data_paths()
     
     for filename in stocks:
-        df = pandas.read_csv(filename)
-        df = df.rename(columns=rename_mapping)
-        df = reverse_df(df)
-        df.to_csv(filename, index=False)
+        if os.path.exists(filename):
+            df = pd.read_csv(filename)
+            df = df.rename(columns=rename_mapping)
+            # df = adjust_for_splits(df)
+            df = reverse_df(df)
+            df.to_csv(filename, index=False)
 
 def install_data(stock_ticker):
     """
@@ -51,7 +65,8 @@ def install_data(stock_ticker):
     :return: None
     """
     file_path = path_to_data + stock_ticker + ".csv"
-    data, meta_data = ts.get_daily(symbol=stock_ticker, outputsize='full')
+    data, meta_data = ts.get_daily_adjusted(symbol=stock_ticker, outputsize='full')
+    # print(data)
     data.to_csv(file_path)
     
 def reverse_df(df):
@@ -71,7 +86,7 @@ def get_data(filename):
     :param filename: path to the file
     :return: pandas dataframe of the file
     """
-    return pandas.read_csv(filename)
+    return pd.read_csv(filename)
 
 def get_all_stocks():
     """
@@ -83,7 +98,9 @@ def get_all_stocks():
     file = open("../data/stocks.txt", 'r')
     l = []
     for stock in file:
-        l.append((stock, get_data(path_to_data + stock.strip() + ".csv")))
+        path = path_to_data + stock.strip() + ".csv"
+        if os.path.exists(path):
+            l.append((stock, get_data(path)))
     return l
 
 def download_all_data():
@@ -92,5 +109,7 @@ def download_all_data():
     """
     file = open("../data/stocks.txt", 'r')
     for l in file:
-        install_data(l.strip())
-        time.sleep(12.1)
+        if l >= "WU":
+            print("Downloaded:", l.strip())
+            install_data(l.strip())
+            time.sleep(12.1)
